--- js/rules/rules-template.js
@@ -1,59 +1,0 @@
-// Экран 'Правила игры'
-
-import {makeDomNode} from '../utils.js';
-
-const PREVIOUS_BUTTON_SELECTOR = `.back`;
-const NAME_INPUT_SELECTOR = `.rules__input`;
-const SEND_BUTTON_SELECTOR = `.rules__button`;
-
-const getTemplate = () => {
-  return `
-    <header class="header">
-      <button class="back">
-        <span class="visually-hidden">Вернуться к началу</span>
-        <svg class="icon" width="45" height="45" viewBox="0 0 45 45" fill="#000000">
-          <use xlink:href="img/sprite.svg#arrow-left"></use>
-        </svg>
-        <svg class="icon" width="101" height="44" viewBox="0 0 101 44" fill="#000000">
-          <use xlink:href="img/sprite.svg#logo-small"></use>
-        </svg>
-      </button>
-    </header>
-    <section class="rules">
-      <h2 class="rules__title">Правила</h2>
-      <ul class="rules__description">
-        <li>Угадай 10 раз для каждого изображения фото
-          <img class="rules__icon" src="img/icon-photo.png" width="32" height="31" alt="Фото"> или рисунок
-          <img class="rules__icon" src="img/icon-paint.png" width="32" height="31" alt="Рисунок"></li>
-        <li>Фотографиями или рисунками могут быть оба изображения.</li>
-        <li>На каждую попытку отводится 30 секунд.</li>
-        <li>Ошибиться можно не более 3 раз.</li>
-      </ul>
-      <p class="rules__ready">Готовы?</p>
-      <form class="rules__form">
-        <input class="rules__input" type="text" placeholder="Ваше Имя">
-        <button class="rules__button  continue" type="submit" disabled>Go!</button>
-      </form>
-    </section>
-  `;
-};
-
-const getTemplateNode = (cbs) => {
-  const eventListeners = [{
-    selector: PREVIOUS_BUTTON_SELECTOR,
-    type: `click`,
-    cb: cbs[0],
-  }, {
-    selector: NAME_INPUT_SELECTOR,
-    type: `keyup`,
-    cb: cbs[1],
-  }, {
-    selector: SEND_BUTTON_SELECTOR,
-    type: `click`,
-    cb: cbs[2],
-  }];
-
-  return makeDomNode(getTemplate(), eventListeners);
-};
-
-export default getTemplateNode;
